Simulate gridded weather function

Description goes here....

Usage

generateWeatherSeries(
  weather.data = NULL,
  weather.grid = NULL,
  weather.date = NULL,
  variable.names = NULL,
  variable.labels = NULL,
  variable.units = NULL,
  sim.year.num = NULL,
  sim.year.start = 2020,
  month.start = 1,
  realization.num = 5,
  warm.variable = "precip",
  warm.signif.level = 0.9,
  warm.sample.num = 5000,
  warm.subset.criteria = NULL,
  knn.sample.num = 100,
  mc.wet.threshold = 0.3,
  mc.extreme.quantile = 0.8,
  evaluate.model = FALSE,
  evaluate.grid.num = 20,
  output.path = getwd(),
  seed = NULL
)

Arguments

weather.data

list of data frames of daily weather observations per grid cell. Each data frame, columns are weather variables and rows are daily values.

weather.grid

Data frame of grid cells. Each grid cell is assigned an id starting from 1, x and y coordinate index value, and x and y coordinates.

weather.date

placeholder

variable.names

vector of names for the variables to be included in the weather generator

variable.labels

vector of labels for the weather variables (optional). If no values provided, it is labels will be same as the names

variable.units

vector of units for each of the weather variables (optional). If no values provided, a blank vector is used.

sim.year.num

numeric value indicating the desired total number of years of simulated weather realizations

sim.year.start

numeric value indicating the starting year of the generated time-series

month.start

the first month of the year (default value is 1). Use a value other than 1 for water-year based analyses

realization.num

number of natural variability realizations to be generated.

warm.variable

the name of the variable for the wavelet autoregressive mode. Default value is precipitation variable.

warm.signif.level

the significance level for the warm model.

warm.sample.num

number of annual sequeces to be generated from the the warm model

warm.subset.criteria

placeholder

knn.sample.num

number of knn years to be sampled

mc.wet.threshold

placeholder

mc.extreme.quantile

placeholder

evaluate.model

logical value indicating wether to save model evaluation plots

evaluate.grid.num

Number of grid cells to be sampled in the evaluation plots

output.path

output path for the weather generator results (string)

seed

placeholder

Value